Structs.
Parse and typecheck structs.

+++ back/datatype/datatype_amd64.go
@@ -1,3 +1,3 @@
 package datatype
 
-const PTR_SIZE byte = 8
+const PTR_SIZE uint32 = 8
