+++ back/datatype/datatype_pointer.go
@@ -8,10 +8,10 @@ func (typ PointerType) Name() string {
 	return "*" + typ.Pointed_type.Name()
 }
 
-func (typ PointerType) BitSize() byte {
+func (typ PointerType) BitSize() uint32 {
 	return PTR_SIZE * 8
 }
 
-func (typ PointerType) ByteSize() byte {
+func (typ PointerType) ByteSize() uint32 {
 	return PTR_SIZE
 }
